relocated findChildrenByMark from focalboard to prosemirror
