+++ lib/inline-comments/findTotalInlineComments.ts
@@ -1,5 +1,5 @@
 import { EditorView, MarkType, Node } from '@bangle.dev/pm';
-import { findChildrenByMark } from 'lib/focalboard/findChildrenByMark';
+import { findChildrenByMark } from 'lib/prosemirror/findChildrenByMark';
 import { ThreadWithCommentsAndAuthors } from 'lib/threads/interfaces';
 
 export function findTotalInlineComments (
